backend/requirements.txt
-i https://pypi.org/simple
annotated-types==0.6.0; python_version >= '3.8'
anyio==4.3.0; python_version >= '3.8'
asttokens==2.4.1
backcall==0.2.0
certifi==2024.2.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
cloudinary==1.39.1
colorama==0.4.6
decorator==5.1.1; python_version >= '3.5'
elevenlabs==1.0.4; python_version >= '3.8' and python_version < '4.0'
exceptiongroup==1.2.0; python_version < '3.11'
executing==2.0.1; python_version >= '3.5'
fastapi==0.110.1; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.5; python_version >= '3.8'
httptools==0.6.1
httpx==0.27.0; python_version >= '3.8'
idna==3.6; python_version >= '3.5'
ipython==8.12.3; python_version >= '3.8'
jedi==0.19.1; python_version >= '3.6'
matplotlib-inline==0.1.6; python_version >= '3.5'
parso==0.8.4; python_version >= '3.6'
pickleshare==0.7.5
prompt-toolkit==3.0.43; python_full_version >= '3.7.0'
pure-eval==0.2.2
pydantic==2.6.4; python_version >= '3.8'
pydantic-core==2.16.3; python_version >= '3.8'
pygments==2.17.2; python_version >= '3.7'
python-dotenv==1.0.1; python_version >= '3.8'
pyyaml==6.0.1
requests==2.31.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
stack-data==0.6.3
starlette==0.37.2; python_version >= '3.8'
traitlets==5.14.2; python_version >= '3.8'
typing-extensions==4.11.0; python_version >= '3.8'
urllib3==2.2.1; python_version >= '3.8'
uvicorn[standard]==0.29.0; python_version >= '3.8'
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0; python_version >= '3.8'
